tests/torture_rekey: Do not mix integer types
In torture_rekey_time(), "long long" was used instead of uint32_t.  This
makes the test to fail in some architectures where the sizes don't match.

+++ tests/client/torture_rekey.c
@@ -327,7 +327,7 @@ static void torture_rekey_time(void **state)
     int rc;
     char data[256];
     unsigned int i;
-    long long time = 3; /* 3 seconds */
+    uint32_t time = 3; /* 3 seconds */
     struct ssh_crypto_struct *c = NULL;
     unsigned char *secret_hash = NULL;
